fixed it coverage, #21 Upgrade to JUnit 5
